Initial refactoring
JIRA: RAIL-4003
